requirements.txt
asgiref==3.10.0
black==23.9.1
click==8.1.7
colorama==0.4.6
Django==5.2.8
django-countries==7.5.1
django-extensions==4.1
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
inflection==0.5.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.8
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==80.9.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
